+++ lib/devise.rb
@@ -524,14 +524,6 @@ module Devise
   def self.deprecator
     @deprecator ||= ActiveSupport::Deprecation.new("5.0", "Devise")
   end
-
-  def self.activerecord51? # :nodoc:
-    deprecator.warn <<-DEPRECATION.strip_heredoc
-      [Devise] `Devise.activerecord51?` is deprecated and will be removed in the next major version.
-      It is a non-public method that's no longer used internally, but that other libraries have been relying on.
-    DEPRECATION
-    defined?(ActiveRecord) && ActiveRecord.gem_version >= Gem::Version.new("5.1.x")
-  end
 end
 
 require 'warden'